Add 2nd recipe "Ribs"
Create ribs.html for the 2nd recipe "Ribs".

In it, follow the same structure as in the 1st
recipe "Brisket", which has the <h1> and the
three subsequent <h2>. Under each heading, include the
same <img>, <p>, <ul>, <ol>, and <a> as in "Brisket".

Also add some indentations to the source code, which
"Brisket" doesn't have, making it easier to read
through in the future.