Update style.css
Import 2 fonts (Lato & Russo One) to replace
the default fonts of the browsers.

Style the elements that will be used across
all the HTML files of this project.

Style the classes that are added to index.html.
These classes are specific to index.html only.
They do not appear on other pages.

Separate each section above with a comment.