Modify 1st recipe "Brisket"
Link style.css to brisket.html inside
the <head>.

Add a class for the main image to set
its height & width. That is to reserve that
portion of the page for the image incase
it loads slower than other contents.

Edit the 'back to homepage' link to make it
cleaner.